refactor: `TokenProvider` 인터페이스 생성 및 `SharedPreferenceTokenProvider` 로 이름 변경
