+++ app/src/main/java/woowacourse/shopping/data/network/AuthorizationInterceptor.kt
@@ -2,7 +2,7 @@ package woowacourse.shopping.data.network
 
 import okhttp3.Interceptor
 import okhttp3.Response
-import woowacourse.shopping.data.TokenProvider
+import woowacourse.shopping.data.token.TokenProvider
 
 class AuthorizationInterceptor(private val tokenProvider: TokenProvider) : Interceptor {
     override fun intercept(chain: Interceptor.Chain): Response {
