+++ app/src/main/java/woowacourse/shopping/data/token/TokenProvider.kt
@@ -1,0 +1,10 @@
+package woowacourse.shopping.data.token
+
+interface TokenProvider {
+    fun getToken(): String?
+
+    fun setToken(
+        name: String,
+        password: String,
+    )
+}
